requirements.txt
appdirs==1.4.4
bandit==1.7.6
blinker==1.7.0
build==1.0.3
buildozer==1.5.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.6
distlib==0.3.8
docutils==0.20.1
dparse==0.6.3
ffpyplayer==4.5.1
filelock==3.13.1
Flask==3.0.0
Flask-MySQL==1.5.2
fonttools==4.47.0
FormEncode==2.1.0
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.2
gunicorn==20.1.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Kivy @ https://github.com/kivy/kivy/archive/stable.zip
Kivy-examples==2.2.1
Kivy-Garden==0.1.5
kivymd==1.1.1
kiwisolver==1.4.5
line-profiler==4.1.2
Mako==1.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
memory-profiler==0.61.0
mysql-connector-python==8.2.0
mysqlclient==2.2.1
numpy==1.26.2
packaging==21.3
pbr==6.0.0
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.1.0
pluggy==1.3.0
plyer==2.1.0
protobuf==4.21.12
psutil==5.9.7
ptyprocess==0.7.0
PyDispatcher==2.0.7
Pygments==2.17.2
pyjnius==1.5.0
pyparsing==3.1.1
pyproject_hooks==1.0.0
pytest==7.4.3
python-dateutil==2.8.2
python-for-android==2023.9.16
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.0
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
safety==2.3.5
sh==1.14.3
six==1.16.0
smmap==5.0.1
SQLObject==3.11.0
stevedore==5.1.0
toml==0.10.2
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
virtualenv==20.25.0
Werkzeug==3.0.1
